% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.pool}
\alias{create.pool}
\title{Simulates a species pool}
\usage{
create.pool(n.species.pool, n.invader.pool, evol.model, min.phyl.signal,
  evol.model.param, nrep = 499)
}
\arguments{
\item{n.species.pool}{number of species (natives and invaders)}

\item{n.invader.pool}{number of invaders}

\item{evol.model}{choice of evolutionary model which determines phylogenetic signal, see details}

\item{min.phyl.signal}{minimum level of phylogenetic signal accepetd, if min.phyl.signal=NA no minimum limit}

\item{evol.model.param}{applies only if evol.model is different from Brownian motion, see details}

\item{nrep}{number of repetition for phylogenetic signal and tree imbalance tests}
}
\value{
List of objects:
  \describe{
  \item{func}{ evolved functional traits for each species of the pool }
  \item{phylo}{ phylogenetic tree of the species pool  }
  \item{indices}{ results of phylogenetic signal and colless test for phylogenetic tree shape }
  \item{invader}{ list of the invader's identities and trait values }
  \item{parameters}{ the parameters used for creating the species pool }
  }
}
\description{
Simulates a species pool containing a phylogeny, species trait values and information on invasion status.
}
\details{
The different evolutionary models (argument: evol.mod) are based on
tree transformations. These are done using the function rescale in
the geiger package and are coded as follows: "1" = BM (Brownian motion),
"2" = deltaTree, "3" = kappaTree, "4" = ouTree, 5 = "lambdaTree".
The tree transformations can be parameterized using the argument
evol.model.param.
}
\examples{
pool <- create.pool(n.species.pool=500, n.invader.pool=1, evol.model="deltaTree", min.phyl.signal=NA, evol.model.param=0.1, nrep=1)
# plot tree
plot(pool$phy)
# calculate phylogenetic signal
phylosignal(pool$func$niche_evol, pool$phy, checkdata=FALSE)
}
\seealso{
\code{\link{simulation.experiment}} for running simulation
experiments with an integrated simulation of species pools and
diversity analyses of final community structures.
}

